Refactor index route - display headlines without search - Changed ‘xmlhttprequest’ package for ‘request’ package - Fixed issue with callback - implemented properly - Remove ‘/users’ route and relevant files
